CVS: ---------------------------------------------------------------------- CVS: Please format log message lines to be no longer than the above line. MSVC++ 2005 compatibility enhanced, clean compile for Release and Debug. Line-ending hell should be over; use LF only from now on.
git-svn-id: file:///desktop/SourceForge%20Backup/sphere/svnrepos/trunk@1272 c0eb2ce6-7f40-480a-a1e0-6c4aea08c2c2

+++ sphere/source/editor/LayerView.cpp
@@ -563,7 +563,7 @@ CLayerView::OnFlattenVisibleLayers()
 
   std::vector<sLayer> layers;
   // get a list of the visible layers
-  for (i = 0; i < m_Map->GetNumLayers(); i++) {
+  for (i = 0; i < (unsigned int)m_Map->GetNumLayers(); i++) {
     if (m_Map->GetLayer(i).IsVisible()) {
       layers.push_back(m_Map->GetLayer(i));
     }
